completed prop drilling and components for list item and items list.
