+++ react-tutorial-playlist/src/components/LineItem.js
@@ -1,0 +1,25 @@
+import { FaTrashAlt } from 'react-icons/fa'
+
+const LineItem = ({ item, handleCheck, handleDelete }) => {
+  return (
+    <li className='item' key={item.id}>
+          <input
+            type='checkbox'
+            checked={item.checked}
+            onChange={() => handleCheck(item.id)}
+          />
+          <label
+            onDoubleClick={() => handleCheck(item.id)}
+            style={(item.checked) ? { textDecoration: 'line-through'} : null}
+          >{item.item}</label>
+          <FaTrashAlt
+            role='button'
+            tabIndex='0'
+            onClick={() => handleDelete(item.id)}
+            aria-label={`Delete ${item.item}`}
+          />
+        </li>
+  )
+}
+
+export default LineItem
