+++ react-tutorial-playlist/src/components/header.js
@@ -1,11 +1,18 @@
+// you can destructure the props in the parameter list so that you
+// don't have to use props then below props.title
 
-const Header = () => {
+const Header = ({ title }) => {
 
   return (
     <header>
-      <h1>Grocery List</h1>
+      <h1>{title}</h1>
     </header>
   )
+}
+
+// example of a default prop
+Header.defaultProps = {
+  title: "Default Title"
 }
 
 export default Header;
